mocks for app type dao moved into separate file (mock_daos.go -> mock_dao_application_type.go)
