+++ internal/testutils/mocks/mock_daos.go
@@ -1,7 +1,6 @@
 package mocks
 
 import (
-	"errors"
 	"fmt"
 	"github.com/RedHatInsights/sources-api-go/config"
 	"github.com/RedHatInsights/sources-api-go/internal/testutils/fixtures"
@@ -10,11 +9,6 @@ import (
 )
 
 var conf = config.Get()
-
-type MockApplicationTypeDao struct {
-	ApplicationTypes []m.ApplicationType
-	Compatible       bool
-}
 
 type MockSourceTypeDao struct {
 	SourceTypes []m.SourceType
@@ -35,21 +29,6 @@ type MockApplicationAuthenticationDao struct {
 
 type MockAuthenticationDao struct {
 	Authentications []m.Authentication
-}
-
-func (a *MockApplicationTypeDao) List(limit int, offset int, filters []util.Filter) ([]m.ApplicationType, int64, error) {
-	count := int64(len(a.ApplicationTypes))
-	return a.ApplicationTypes, count, nil
-}
-
-func (a *MockApplicationTypeDao) GetById(id *int64) (*m.ApplicationType, error) {
-	for _, i := range a.ApplicationTypes {
-		if i.Id == *id {
-			return &i, nil
-		}
-	}
-
-	return nil, util.NewErrNotFound("application type")
 }
 
 func (a *MockMetaDataDao) List(limit int, offset int, filters []util.Filter) ([]m.MetaData, int64, error) {
@@ -110,80 +89,6 @@ func (md *MockMetaDataDao) GetSuperKeyAccountNumber(applicationTypeId int64) (st
 
 func (md *MockMetaDataDao) ApplicationOptedIntoRetry(applicationTypeId int64) (bool, error) {
 	panic("not implemented!")
-}
-
-func (a *MockApplicationTypeDao) Create(src *m.ApplicationType) error {
-	panic("not implemented") // TODO: Implement
-}
-
-func (a *MockApplicationTypeDao) Update(src *m.ApplicationType) error {
-	panic("not implemented") // TODO: Implement
-}
-
-func (a *MockApplicationTypeDao) Delete(id *int64) error {
-	panic("not implemented") // TODO: Implement
-}
-
-func (a *MockApplicationTypeDao) SubCollectionList(primaryCollection interface{}, limit, offset int, filters []util.Filter) ([]m.ApplicationType, int64, error) {
-	var appTypesOut []m.ApplicationType
-
-	switch object := primaryCollection.(type) {
-	case m.Source:
-		var sourceExists bool
-		for _, src := range fixtures.TestSourceData {
-			if src.ID == object.ID {
-				sourceExists = true
-			}
-		}
-		if !sourceExists {
-			return nil, 0, util.NewErrNotFound("source")
-		}
-
-		appTypes := make(map[int64]int)
-		for _, app := range fixtures.TestApplicationData {
-			if app.SourceID == object.ID {
-				appTypes[app.ApplicationTypeID]++
-			}
-		}
-
-		for _, appType := range a.ApplicationTypes {
-			for id := range appTypes {
-				if appType.Id == id {
-					appTypesOut = append(appTypesOut, appType)
-					break
-				}
-			}
-		}
-	default:
-		return nil, 0, fmt.Errorf("unexpected primary collection type")
-	}
-
-	count := int64(len(appTypesOut))
-	return appTypesOut, count, nil
-}
-
-func (a *MockApplicationTypeDao) ApplicationTypeCompatibleWithSource(_, _ int64) error {
-	if a.Compatible {
-		return nil
-	}
-
-	return errors.New("Not compatible!")
-}
-
-func (at *MockApplicationTypeDao) GetSuperKeyResultType(applicationTypeId int64, authType string) (string, error) {
-	panic("not needed")
-}
-
-func (a *MockApplicationTypeDao) ApplicationTypeCompatibleWithSourceType(_, _ int64) error {
-	if a.Compatible {
-		return nil
-	}
-
-	return errors.New("Not compatible!")
-}
-
-func (a *MockApplicationTypeDao) GetByName(_ string) (*m.ApplicationType, error) {
-	return nil, nil
 }
 
 func (a *MockSourceTypeDao) List(limit int, offset int, filters []util.Filter) ([]m.SourceType, int64, error) {
